requirements.txt
python-telegram-bot[ext]>=20.7
mutagen>=1.47.0
python-dotenv>=1.0.0
